[fix] sks ta: fix handle database against spectre issue
Implement index sanitation is assembly, mandatory at least for the
sanitation mask(s) generation.

Update handle API description: 0 denote invalid returned handles.

+++ ta_services/secure_key_services/src/handle.h
@@ -8,10 +8,11 @@
 #define __HANDLE_H
 
 #include <stddef.h>
+#include <unistd.h>
 
 struct handle_db {
 	void **ptrs;
-	uint32_t max_ptrs;
+	ssize_t max_ptrs;
 };
 
 #define HANDLE_DB_INITIALIZER { NULL, 0 }
@@ -27,8 +28,8 @@ void handle_db_destroy(struct handle_db *db);
  * Allocates a new handle and assigns the supplied pointer to it,
  * ptr must not be NULL.
  * The function returns
- * >= 0 on success and
- * -1 on failure
+ * > 0 on success and
+ * 0 on failure
  */
 uint32_t handle_get(struct handle_db *db, void *ptr);
